fix: some code refactoring
