+++ src/main/java/com/github/daniilandco/vehicle_sales_project/database_access/user/user_model/Permission.java
@@ -1,8 +1,8 @@
 package com.github.daniilandco.vehicle_sales_project.database_access.user.user_model;
 
 public enum Permission {
-    USERS_READ("users:read"),
-    USERS_WRITE("users:write");
+    MANAGE("admins:manage"),   // These permissions are temp. so far : in progress
+    USE("users:use");
 
     private final String permission;
