+++ src/main/java/com/github/daniilandco/vehicle_sales_project/rest/auth/login/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.github.daniilandco.vehicle_sales_project.rest.login;
+package com.github.daniilandco.vehicle_sales_project.rest.auth.login;
 
 import lombok.Data;
