--- src/main/java/com/github/daniilandco/vehicle_sales_project/rest/register/RegisterResponse.java
@@ -1,11 +1,0 @@
-package com.github.daniilandco.vehicle_sales_project.rest.register;
-
-import lombok.Data;
-
-@Data
-public class RegisterResponse {
-    private String message;
-    public RegisterResponse(String message) {
-        this.message = message;
-    }
-}
