+++ src/main/resources/application.properties
@@ -8,4 +8,3 @@ spring.datasource.driver-class-name =com.mysql.jdbc.Driver
 jwt.header=Authorization
 jwt.secret=ItechArt-internship-2021-Daniel-Bondarkov
 jwt.expiration=150803
-jwt.encryptionStrength = 12
